Force le 32bits sur la gestion des adresse IP
git-svn-id: svn://localhost/qrezix/branches/qrezix--qt4@826 30024bfc-89fa-0310-95ff-ee513f6fc940

+++ qrezix/rzxconfig.h
@@ -227,7 +227,7 @@ public:
 	static const QPixmap& localhostIcon();
 	static void saveIcon(const QString& name, const QPixmap& image);
 	
-	static QString historique(unsigned long ip, const QString& hostname);
+	static QString historique(quint32 ip, const QString& hostname);
 	
 	static QString buildLocalhost();
